:white_check_mark: (test): Added more Jest tests
